+++ Sources/RichTextKit/Images/Image+RichText.swift
@@ -26,7 +26,9 @@ public extension Image {
     
     static let richTextColorBackground = symbol("highlighter")
     static let richTextColorForeground = symbol("character")
-
+    static let richTextColorStroke = symbol("a.square")
+    static let richTextColorStrikethrough = symbol("strikethrough")
+    
     static let richTextDocument = symbol("doc.text")
     static let richTextDocuments = symbol("doc.on.doc")