+++ Sources/RichTextKit/Styles/RichTextStyleButton.swift
@@ -61,14 +61,7 @@ public struct RichTextStyleButton: View {
         self.init(
             style: style,
             buttonStyle: buttonStyle,
-            value: {
-                switch style {
-                case .bold: return context.isBoldBinding
-                case .italic: return context.isItalicBinding
-                case .strikethrough: return context.isStrikethroughBinding
-                case .underlined: return context.isUnderlinedBinding
-                }
-            }(),
+            value: context.binding(for: style),
             fillVertically: fillVertically
         )
     }