+++ RELEASE_NOTES.md
@@ -14,7 +14,9 @@ This version also consolidates all separate attribute readers and writers into t
 
 ### ✨ Features
 
+* `RichTextColor` is a new enum that defines supported colors that can be set.
 * `RichTextColorPicker` now works on all platforms.
+* `RichTextContext` has new enum-based style and color functions.
 * `RichTextAttributeReader` can handle more colors and superscripting.
 * `RichTextAttributeWriter` can handle more colors and superscripting.